# sss-modernization-demo/backend/API-SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: SSS Modernization Platform API
  description: Enterprise exemption management and case processing API
  version: 1.0.0
  contact:
    name: SSS Platform Team
    email: platform@sss-modernization.gov
  license:
    name: Government Use Only
servers:
  - url: https://api.sss-modernization.example.com/api
    description: Production API
  - url: https://staging-api.sss-modernization.example.com/api
    description: Staging API
  - url: http://localhost:5000/api
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with 1-hour TTL
  schemas:
...
    Case:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum: [Draft, Submitted, "In Review", Approved, Denied, Appealed]
        type:
          type: string
          example: "Exemption Request"
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, status, type, reason, createdAt, updatedAt]
...
paths:
  /auth/register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                fullName:
                  type: string
              required: [email, password, fullName]
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  fullName:
                    type: string
        '400':
          description: Invalid input
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: Authenticate user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT bearer token
                  expiresIn:
                    type: integer
                    description: Token TTL in seconds
                  mfaRequired:
                    type: boolean
        '401':
          description: Invalid credentials
...
  /cases:
    get:
      summary: List cases
      tags: [Cases]
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by case status
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Page number (default 1)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Results per page (default 20)
      responses:
        '200':
          description: List of cases
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Case'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
        '401':
          description: Unauthorized

    post:
      summary: Create new case
      tags: [Cases]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Draft, Submitted]
                type:
                  type: string
                reason:
                  type: string
              required: [status, type, reason]
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
...
security:
  - BearerAuth: []
...